ENH: Modularize files
Modularize files:
- Make the reusable methods dwell in the `brainmatch.py` module.
- Make `brainmatch` be a module by adding its `__init__.py` file.
- Make the script computing the brainmatch scores dwell in a separate module.

Change the call to the script in the `README` accordingly.
